+++ ui.apps/src/main/content/jcr_root/apps/aemlab/clientlibs/clientlib-site/header.css
@@ -1,14 +1,7 @@
-header.experiencefragment {
+#aemlab-header {
 	padding:0 6rem;
 	background : #3D4750;
 	margin-bottom: 2rem;
-}
-
-header .cmp-container>.text{
-    padding:0;
-}
-
-header .cmp-experiencefragment--header .cmp-container{
     flex: 1 1 auto;
     position: relative;
     width: auto;
@@ -16,11 +9,17 @@ header .cmp-experiencefragment--header .cmp-container{
     align-items: stretch;
 }
 
-header .cmp-experiencefragment--header .cmp-containera:after{
+#aemlab-header >.text{
+    padding:0;
+}
+
+
+
+#aemlab-header .cmp-containera:after{
 	content:'';
 }
 
-header .cmp-experiencefragment--header #header-logo{
+#aemlab-header #header-logo{
     flex: 1 1 auto;
     position: relative;
     width: auto;
@@ -28,13 +27,13 @@ header .cmp-experiencefragment--header #header-logo{
     margin:1rem 1rem 0 0;
 }
 
-header .cmp-experiencefragment--header #header-logo p, header .cmp-experiencefragment--header #header-logo p a{
+#aemlab-header #header-logo p, #aemlab-header #header-logo p a{
     font-size:2rem;
     color:#026491;
     font-family:CustomFont;
 }
 
-header .cmp-experiencefragment--header #header-logo p+p, header .cmp-experiencefragment--header #header-logo p+p a{
+#aemlab-header #header-logo p+p, #aemlab-header #header-logo p+p a{
     color:#E5710B;
 }
 
@@ -58,20 +57,20 @@ header .cmp-experiencefragment--header #header-logo p+p, header .cmp-experiencef
     box-shadow: 0 0.3125rem 0.9375rem 0 rgb(231 68 68 / 25%);
 }
 
-header .cmp-experiencefragment--header #header-navigation ul {
+#aemlab-header #header-navigation ul {
     position: relative;
     display: flex;
     margin: 1rem 2rem 0 30rem;
     overflow: hidden;
 }
 
-header .cmp-experiencefragment--header #header-navigation ul li{
+#aemlab-header #header-navigation ul li{
     list-style:none;
     padding:0 0.75rem;
     font-size:1rem;
 }
 
-header .cmp-experiencefragment--header a:after{
+#aemlab-header a:after{
 	content:'';
 }
