+++ ui.content/src/main/content/jcr_root/conf/aemlab/_sling_configs/com.community.aemlab.core.conf.HeaderFooterConfig/.content.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
+    jcr:primaryType="sling:Folder"
+    sling:configPropertyInherit="{Boolean}true"
+    footerPath="/content/experience-fragments/aemlab/us/en/site/footer/master"
+    headerPath="/content/experience-fragments/aemlab/us/en/site/header/master"/>
