--- ui.content/src/main/content/jcr_root/conf/aemlab/_sling_configs/template/headerfooter/.content.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root 
-    xmlns:sling="http://sling.apache.org/jcr/sling/1.0"
-    xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="cq:Page">
-    <jcr:content jcr:primaryType="cq:PageContent"
-    			 sling:configPropertyInherit="{Boolean}true"
-    			 footerPath="/content/experience-fragments/aemlab/language-masters/en/site/footer/master"
-    			 headerPath="/content/experience-fragments/aemlab/language-masters/en/site/header/master"
-    />
-</jcr:root>
