+++ ui.content/src/main/content/jcr_root/conf/aemlab/settings/wcm/templates/page-content-cac/structure/.content.xml
@@ -13,7 +13,7 @@
             <header
                 jcr:primaryType="nt:unstructured"
                 sling:resourceType="aemlab/concept/components/cac-xf"
-                configName="template/header"
+                configName="com.community.aemlab.core.conf.HeaderFooterConfig"
                 property="headerPath"
                 noConfigLabel="No configuration defined for Header"/>
             <container
@@ -28,7 +28,7 @@
             <footer
                 jcr:primaryType="nt:unstructured"
                 sling:resourceType="aemlab/concept/components/cac-xf"
-                configName="template/footer"
+                configName="com.community.aemlab.core.conf.HeaderFooterConfig"
                 property="footerPath"
                 noConfigLabel="No configuration defined for Footer"/>
         </root>
