+++ ui.content/src/main/content/jcr_root/content/experience-fragments/aemlab/us/en/site/header/master/.content.xml
@@ -2,6 +2,8 @@
 <jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0"
     jcr:primaryType="cq:Page">
     <jcr:content
+        cq:lastModified="{Date}2022-08-28T19:11:08.633+02:00"
+        cq:lastModifiedBy="admin"
         cq:tags="[]"
         cq:template="/conf/aemlab/settings/wcm/templates/xf-web-variation"
         cq:xfMasterVariation="{Boolean}true"
@@ -13,23 +15,33 @@
             jcr:primaryType="nt:unstructured"
             sling:resourceType="aemlab/components/container"
             layout="simple">
-            <text
+            <container
+                jcr:created="{Date}2022-08-28T19:08:06.669+02:00"
+                jcr:createdBy="admin"
+                jcr:lastModified="{Date}2022-08-28T19:11:08.558+02:00"
+                jcr:lastModifiedBy="admin"
                 jcr:primaryType="nt:unstructured"
-                sling:resourceType="aemlab/components/text"
-                id="header-logo"
-                text="&lt;p>&lt;a href=&quot;#&quot;>AEM&lt;/a>&lt;/p>&#xd;&#xa;&lt;p>&lt;a href=&quot;#&quot;>Lab&lt;/a>&lt;/p>&#xd;&#xa;"
-                textIsRich="true"/>
-            <text_628130385
-                jcr:primaryType="nt:unstructured"
-                sling:resourceType="aemlab/components/text"
-                id="header-navigation"
-                text="&lt;ul>&#xd;&#xa;&lt;li>&lt;a href=&quot;#&quot;>Services&lt;/a>&lt;/li>&#xd;&#xa;&lt;li>&lt;a href=&quot;#&quot;>About Us&lt;/a>&lt;/li>&#xd;&#xa;&lt;li>&lt;a href=&quot;#&quot;>Contact Us&lt;/a>&lt;/li>&#xd;&#xa;&lt;li>&lt;a href=&quot;#&quot;>Careers&lt;/a>&lt;/li>&#xd;&#xa;&lt;/ul>&#xd;&#xa;"
-                textIsRich="true"/>
-            <search
-                jcr:primaryType="nt:unstructured"
-                sling:resourceType="aemlab/components/search"
-                id="header-search"
-                searchRoot="/content/aemlab/reference-content"/>
+                sling:resourceType="aemlab/components/container"
+                id="aemlab-header"
+                layout="simple">
+                <text_copy
+                    jcr:primaryType="nt:unstructured"
+                    sling:resourceType="aemlab/components/text"
+                    id="header-logo"
+                    text="&lt;p>&lt;a href=&quot;#&quot;>AEM&lt;/a>&lt;/p>&#xd;&#xa;&lt;p>&lt;a href=&quot;#&quot;>Lab&lt;/a>&lt;/p>&#xd;&#xa;"
+                    textIsRich="true"/>
+                <text_628130385_copy
+                    jcr:primaryType="nt:unstructured"
+                    sling:resourceType="aemlab/components/text"
+                    id="header-navigation"
+                    text="&lt;ul>&#xd;&#xa;&lt;li>&lt;a href=&quot;#&quot;>Services&lt;/a>&lt;/li>&#xd;&#xa;&lt;li>&lt;a href=&quot;#&quot;>About Us&lt;/a>&lt;/li>&#xd;&#xa;&lt;li>&lt;a href=&quot;#&quot;>Contact Us&lt;/a>&lt;/li>&#xd;&#xa;&lt;li>&lt;a href=&quot;#&quot;>Careers&lt;/a>&lt;/li>&#xd;&#xa;&lt;/ul>&#xd;&#xa;"
+                    textIsRich="true"/>
+                <search
+                    jcr:primaryType="nt:unstructured"
+                    sling:resourceType="aemlab/components/search"
+                    id="header-search"
+                    searchRoot="/content/aemlab/reference-content"/>
+            </container>
         </root>
     </jcr:content>
 </jcr:root>
